Add hello world codegen example
This PR adds a codegen inference example for the hello world model to
demonstrate how to invoke the code generator and build the generated
source. For now, we're just checking the generated source into the repo
to skip over building out the make rules and also ensuring the generated
source complies with formatting rules.

This also fixes a minor formatting issue in the source templates, as
clang-format now properly complained about it.

BUG=b/295390000