Break dependency of kernels on MicroGraph
The control flow and resource kernels make calls into MicroGraph by
using the MicroContext::graph() accessor. For interpreter-less invoke,
we don't need a full MicroGraph, only the subset of functionality that
the kernels use. This PR introduces an abstract base class for
MicroGraphInfo, which MicroGraph implements. The kernels only depend on
MicroGraphInfo, but the MicroContext will still have a full MicroGraph.

BUG=b/295174086