Fix RISC-V codegen example
The RISC-V toolchain we're using failed to properly discern the
TfLiteEvalTensor type for use with an assignment operator. It produced
compiler errors for "no match for 'operator=' with operand types
TfLiteEvalTensor and a brace-closed initializer list. This PR resolves
this issue by explicitly using the TfLiteEvalTensor constructor for the
initializer list. We also apply this approach to the TfLiteNode
initialization as well, just for consistency.

BUG=#2195