Define preprocessor schema
The purpose of the preprocessor is to load a model with the TFLM
interpreter, allocate tensors, and then capture the resulting data
structures. To capture this data, we need to provide structure and
serialization to it for ingesting into the code generator. This commit
defines a preprocessor data schema with generated C++ and python
bindings for it. Unfortunately, we need to check-in the generated code,
as the Makefiles are currently not capable of running the flatc
compiler. An update script is included to help keep these files in sync.

BUG=b/295076067