Create codegen_preprocessor
This PR creates the initial scaffolding for the codegen preprocessor.
The preprocessor is a target-specific binary that will take a model,
load it into a TFLM Interpreter, perform the Init & Prepare stages, then
serialize the resulting data structures to an output file. Currently,
all this binary does is load the model file and write an output file
that simply contains the source model path in it. This will be expanded
as we expose the data.

BUG=b/295076056