Add Bazel to Xtensa docker image
This commit adds the Xtensa Xplorer install scripts to the repo and
updates their respective Dockerfiles to use the repo's path. Most
importantly, it also adds Bazel to the docker images.

This also necessitated removing the use of sudo from the Bazel install
script, as there is no sudo in the Docker images. This should be okay,
as it is reasonable for users to specify sudo when calling an install
script.