Add Xtensa dockerfiles (#2223)
This PR adds the existing Xtensa dockerfiles and install scripts to the repo. Notably, this does not include any of the install packages themselves.

An additional change on top of the existing dockerfiles was to have Bazel installed in each image. This is necessary for using Bazel as part of the codegen process. Commit 6d0d214 has the original dockerfiles as they existed prior.

The install_bazelisk.sh script was also slightly tweaked to be compatible with docker as well.

BUG=cleanup