Update python version in Xtensa images
The Xtensa docker images were using python 3.6, which is incompatible
with some of our python packages. This has not been a problem before
because the Xtensa images only perform make builds with very minimal
python script usage. With the new codegen tools, Bazel is used to invoke
the code generator, and this brings in the incompatible python
dependencies.

This PR changes the base image for the three Xtensa docker containers to
pull to python instead of ubuntu. This is similar to our base TFLM
docker image, which also uses the same python:3.10-bullseye base image.

BUG=b/300655634