Initialize ArithmeticParams & OpDataReduce
There are several structs that are allocated as either part of the arena
or on the stack that do not get entirely initialized. It is typical that
individual operations will just fill in the fields of a struct that they
need. This can lead to difficult to find bugs where a field is used
despite never being initialized. This PR resolves two such instances
found in issue #2336.

The first is in the reference ADD kernel, where we use ArithmeticParams
without initializing it. This is resolved by zero initializing the
struct when it is declared.

The second is the OpDataReduce object that is created in the Persistent
Arena for Reduce kernels. This is resolved by performing placement new
on the buffer that is returned from AllocatePersistentBuffer. This is a
pattern we use elsewhere with allocations, and likely one we should
re-use in all Init functions.

BUG=#2336