Add DISCARD_FUN_FOR_NONVOID_RETURN to ndsp-hifi5
The xa_nnlib_hifi5 library has a modified copy of the NDSP library
contained within it. The TFLM repository has recently switched to using
the standalone copy of NDSP library for all HiFis. There are some
mismatches between these two versions which caused compilation failures
in TFLM for HiFi5. In particular, the DISCARD_FUN_FOR_NONVOID_RETURN
macro was missing from the NDSP HiFi5 library, which is necessary for
compiling xa_nn_matumul_f16.c. This commit adds that macro to the local
TFLM patch that we maintain.