codegen: Add Bazel macro for inference library
The TFLM code_generator takes a TFLite model and generates C/C++ source
code that can then be compiled into a binary. This change adds a Bazel
macro for invoking the code generator and creating a cc_library with the
resulting sources. It also updates the checked-in hello_world example
with an appropriate BUILD file and updates the script to use Bazel
instead of make.

BUG=b/294230402