Add missing ops to OpResolvers
The PythonOpsResolver and the utility TflmOpResolver are intended to
support all built-in ops allow for models to be tested without code
changes. This PR syncs those op resolvers with all available ops from
the MicroMutableOpResolver, notably adding BatchMatMul and the Signal
ops.

Additionaly, this PR sorts the list alphabetically for readability and
adds an alias for the utility TflmOpResolver since it is used in both
the benchmarking tool and the layer by layer debugging tool.

BUG=cleanup