Add meta-data to generic benchmark (#2496)
@tensorflow/micro

Add meta-data strings to the binary.  Meta-data is output each time binary is executed.  See the README for a sample of the output.

If a Git repo. is not available at ```${TENSORFLOW_ROOT}```, an appropriate message is generated in the meta-data.

Changes to how the CMSISNN and XTENSA NN library download scripts work.

Adds run of generic benchmark with embedded model for default build of x86, Corstone-300, Xtensa to test scripts executed during CI.

bug=fixes #2495