feat(compression): implement tensor decompression in op fully_connected
Implement tensor decompression in op fully_connected. Extend
tests to validate operation on compressed tensors.

Fix the effect on memory_arena_threshold_test by setting a
different expected value for the persistent buffer allocation
when compression is configured in. The allocation was allowed to
vary by 3%; however, compression adds ~10%. Set the expected
value to the measured value when compression is configured in.

BUG=part of #2636