build(codegen): suppress noise in console output
Redirect output from the code_generator binary so that when it's
used within the build system, it doesn't print unexpected,
distracting noise to the console. Generally, compiler or
generator commands don't print output unless there's an error.

This reverts "build(codegen): suppress noise in console output
(#2708)", commit d2495773, which attempted to fix the same
problem by adding a --quiet flag. A subsequent upgrade to the
Tensorflow Python package caused new noise that wasn't possible
to suppress with the flag.

BUG=description