fix(create_tflm_tree): remove recent tests from exported tree (#2751)
Remove the recently added span_test.cc and static_vector_test.cc from
the files exported by the create_tflm_tree.py project generation process
by adding them to the list of tests in the Makefile. Unit tests are not
meant to be included in exported trees; they may include files that are
not exported.

This change also ensures that these tests are included when `make test`
is run.

BUG=fixes #2718