fix: restore script called by deleted job
Restore the script called by the deleted job. A PR runs from the
workflows in the commit onto which the PR would be merged, not
the workflows in this current commit. Therefore a job and it's
script cannot be removed at the same time. First, the removal of
the job must be merged. Then, the dangling script can be removed
in a subsequent commit.