build(bazel): provide means of excluding whl_test locally
Mark whl_test as size "large", so that it, and other tests of a
similar size, can be excluded with the --test_size_filters
option.

For even more convenience, provide a hook for developers to add
localy .bazlerc-style defaults and configurations in
$root/bazelrc.local and ../$root/bazelrc.local.

This gives a convenient means of explicitly excluding whl_test
during local, incremental development runs of `bazel test ...`,
while not surprising users who don't opt in, and without
affecting CI.

whl_test's cached result is invalidated, and the test rerun,
following most git activity, because the git hash appears in the
Python package's version number. This combined with its long
runtime and network dependency makes it a nuisance when running
`bazel test ...` incrementally during development. Of course,
this test should be allowed to run when developing changes that
affect the package, before pushing commits to main, and in CI.